Add updated bookstore structure
