+++ src/components/Categories.js
@@ -1,25 +1,5 @@
 import React from 'react';
-import { useDispatch, useSelector } from 'react-redux';
-import { checkStatus } from '../redux/categories/categories';
 
-const Categories = () => {
-  const categories = useSelector((state) => state.categories);
-  const dispatch = useDispatch();
+const Categories = () => (<button type="button">Check status</button>);
 
-  const handleCheckStatus = () => {
-    dispatch(checkStatus());
-  };
-
-  return (
-    <div>
-      {categories.length > 0 && (
-        <h1>{categories}</h1>
-      )}
-      <button type="button" onClick={handleCheckStatus}>
-        Check Status
-      </button>
-    </div>
-  );
-};
-
-export default Categories;
+export default Categories
